refactor: extract terminal and emulator cleanup logic
